Game operational & code refactoring
It works! The game now can play through with all objects now able to be picked up, taken with you, and dropped and the player can now win the game!

As well as the gameplay being finished (likely to be improved upon) I have also organised all methods into their own files and called upon by reference when they're needed to run. I chose to do this to make the code itself easier to read.

Enjoy the first full version!