Added text-to-speech (and minor changes)
The main purpose of this update was to add text-to-speech to all text outputted by the program. This allowed the user to hear what is happening as well as reading it, which I believe makes it more enjoyable to play through.

This was achieved through the use of the System.Speech.Synthesis resources.

The controls are now displayed after the player enters their name as well as changes to the area descriptions (as some had been poorly written which became more apparent once using the TTS).